# backend/offertracker-openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "OfferTracker API",
    "description": "求职管理系统后端API",
    "version": "1.0.0",
    "contact": {
      "name": "OfferTracker Team",
      "url": "https://github.com/offertracker"
    }
  },
  "servers": [
    {
      "url": "https://offertracker-backend.1544718463.workers.dev",
      "description": "生产环境"
    },
    {
      "url": "http://localhost:8787",
      "description": "开发环境"
    }
  ],
...
  "paths": {
...
    "/api/dashboard/summary": {
      "get": {
        "summary": "获取Dashboard汇总数据",
        "description": "获取Dashboard的统计汇总数据",
        "tags": ["数据统计"],
        "security": [{"BearerAuth": []}],
        "responses": {
          "200": {
            "description": "成功获取汇总数据",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardSummary"
                }
              }
            }
          }
        }
      }
    },
    "/api/dashboard/timeline": {
      "get": {
        "summary": "获取时间线数据",
        "description": "获取最近50条求职记录的时间线数据",
        "tags": ["数据统计"],
        "security": [{"BearerAuth": []}],
        "responses": {
          "200": {
            "description": "成功获取时间线数据",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Application"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/dashboard/insights": {
      "get": {
        "summary": "获取洞察数据",
        "description": "获取高级分析洞察数据",
        "tags": ["数据统计"],
        "security": [{"BearerAuth": []}],
        "responses": {
          "200": {
            "description": "成功获取洞察数据",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "monthly_trend": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "month": {
                            "type": "string",
                            "example": "2024-01"
                          },
                          "count": {
                            "type": "integer",
                            "example": 10
                          }
                        }
                      }
                    },
                    "status_processing_time": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "status": {
                            "type": "string",
                            "example": "applied"
                          },
                          "avg_days": {
                            "type": "number",
                            "example": 7.5
                          }
                        }
                      }
                    },
                    "channel_success_rate": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "channel": {
                            "type": "string",
                            "example": "boss"
                          },
                          "success_rate": {
                            "type": "number",
                            "example": 25.0
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/enums": {
      "get": {
        "summary": "获取所有枚举类型",
        "description": "获取系统中所有的枚举类型",
        "tags": ["枚举配置"],
        "security": [{"BearerAuth": []}],
        "responses": {
          "200": {
            "description": "成功获取枚举类型",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "example": "status"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/enums/{type}": {
      "get": {
        "summary": "获取枚举值",
        "description": "获取指定类型的枚举值列表",
        "tags": ["枚举配置"],
        "security": [{"BearerAuth": []}],
        "parameters": [
          {
            "name": "type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "enum": ["status", "city", "salary_range", "channel"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取枚举值",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EnumItem"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/enums/batch": {
      "post": {
        "summary": "批量获取枚举值",
        "description": "批量获取多个类型的枚举值",
        "tags": ["枚举配置"],
        "security": [{"BearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["types"],
                "properties": {
                  "types": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "enum": ["status", "city", "salary_range", "channel"]
                    },
                    "example": ["status", "city"]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功获取批量枚举值",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/EnumItem"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Application": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "application-id"
          },
          "user_id": {
            "type": "string",
            "example": "user-id"
          },
          "resume_id": {
            "type": "string",
            "example": "resume-id"
          },
          "company_name": {
            "type": "string",
            "example": "公司名称"
          },
          "position_title": {
            "type": "string",
            "example": "职位名称"
          },
          "status": {
            "type": "string",
            "example": "applied"
          },
          "city": {
            "type": "string",
            "example": "北京"
          },
          "salary_range": {
            "type": "string",
            "example": "20k-30k"
          },
          "channel": {
            "type": "string",
            "example": "boss"
          },
          "contact_name": {
            "type": "string",
            "example": "联系人"
          },
          "contact_email": {
            "type": "string",
            "format": "email",
            "example": "contact@example.com"
          },
          "application_date": {
            "type": "integer",
            "example": 1234567890
          },
          "last_update": {
            "type": "integer",
            "example": 1234567890
          },
          "interview_date": {
            "type": "integer",
            "example": 1234567890
          },
          "offer_status": {
            "type": "string",
            "example": "accepted"
          },
          "rejection_reason": {
            "type": "string",
            "example": "拒绝原因"
          },
          "notes": {
            "type": "string",
            "example": "备注信息"
          },
          "custom_fields": {
            "type": "string",
            "example": "{}"
          },
          "resume_name": {
            "type": "string",
            "example": "简历名称"
          }
        }
      },
...
      "DashboardSummary": {
        "type": "object",
        "properties": {
          "status_distribution": {
            "type": "object",
            "additionalProperties": {
              "type": "integer"
            },
            "example": {
              "applied": 10,
              "interviewing": 5,
              "offer": 2,
              "rejected": 3
            }
          },
          "city_distribution": {
            "type": "object",
            "additionalProperties": {
              "type": "integer"
            },
            "example": {
              "北京": 8,
              "上海": 6,
              "深圳": 4
            }
          },
          "channel_distribution": {
            "type": "object",
            "additionalProperties": {
              "type": "integer"
            },
            "example": {
              "boss": 12,
              "lagou": 5,
              "internal": 3
            }
          },
          "total_applications": {
            "type": "integer",
            "example": 20
          },
          "success_rate": {
            "type": "number",
            "example": 10.0
          },
          "average_response_time": {
            "type": "number",
            "example": 7.5
          }
        }
      },
      "EnumItem": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "status"
          },
          "value": {
            "type": "string",
            "example": "applied"
          },
          "label": {
            "type": "string",
            "example": "已投递"
          }
        }
      }
    }
  }
}